// generated/src/app/api-models/stage-membr-support.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageMembrSupport {
  batchId: string;
  transactionId: number;
  seqGroupId: number;
  seqProvId: number;
  providerId: string;
  provider2Id: string;
  seqMembId: number;
  seqSubsId: number;
  seqEligHist: number;
  seqEnrollmentRule: number;
  facProvId: string;
  facSeqProvId: number;
}